# cerebral-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cerebral API",
        "description": "API de gestion de projets et tâches pour Cerebral",
        "contact": {
            "name": "Support Cerebral",
            "email": "support@cerebral.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/tasks/{task}/assign": {
            "put": {
                "tags": [
                    "Tâches"
                ],
                "summary": "Assigner une tâche à un utilisateur",
                "operationId": "14556b8c95befa54a6d43f1c013c4cb1",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tâche",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigned_to"
                                ],
                                "properties": {
                                    "assigned_to": {
                                        "description": "ID de l'utilisateur à qui assigner la tâche",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assignment_notes": {
                                        "description": "Notes optionnelles pour l'assignation",
                                        "type": "string",
                                        "example": "Notes d'assignation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tâche assignée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tâche assignée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}